--- samples/Functions-ReplyToDiscussions/HttpTrigger1.cs
@@ -1,26 +1,0 @@
-using System;
-using System.IO;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
-
-namespace MyAmbassadorDemo.Function
-{
-    public static class HttpTrigger1
-    {
-        [FunctionName("HttpTrigger1")]
-        public static async Task<IActionResult> Run(
-            [HttpTrigger(AuthorizationLevel.Function, "post", Route = null)] HttpRequest req,
-            ILogger log)
-        {
-            log.LogInformation("C# HTTP trigger function processed a request.");
-            var job = new RespondToNewComments();
-            await job.RunAsync();
-            return new OkObjectResult("OK");
-        }
-    }
-}
