+++ samples/functions/replyToNewGithubDiscussion/ReplyToNewGithubDiscussion.cs
@@ -2,16 +2,16 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MyAmbassadorDemo.Function
+namespace APIMAuthorizationsDemo.Function
 {
-  public class RespondToNewComments {
+  public class ReplyToNewGithubDiscussion {
     private const string APIM_SERVICE_NAME = "{APIM_SERVICE_NAME}";
     private const string SUBSCRIPTION_KEY = "{SUBSCRIPTION_KEY}";
-    private const string DISCUSSION_ID = "D_kwDOHiNKns4AP3wj";
+    private const string DISCUSSION_ID = "{DISCUSSION_ID}";
 
     private ApimService _apimService;
 
-    public RespondToNewComments() {
+    public ReplyToNewGithubDiscussion() {
       _apimService = new ApimService($"https://{APIM_SERVICE_NAME}.azure-api.net", $"{SUBSCRIPTION_KEY}");
     }
